Create model unit tests and renamed test package
