+++ src/test/java/fetchrewards/userpointstests/UserPointsApplicationTests.java
@@ -1,4 +1,4 @@
-package FetchRewards.UserPoints;
+package fetchrewards.userpointstests;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
